Make `hack/tools` a module and move generators out
