+++ .github/workflows/release.yaml
@@ -35,7 +35,7 @@ jobs:
       github-tag-template: '${version}'
       github-additional-tag-templates: |-
         pkg/apis/${version}
-        hack/tools/tool/${version}
+        hack/tools/${version}
       next-version: ${{ inputs.next-version }}
       slack-channel-id: C9CEBQPGE # #sap-tech-gardener
       assets: |
