+++ .gitignore
@@ -27,8 +27,8 @@ config.vmdk
 /test/framework/resources/repository/cache
 /test/framework/resources/shoots
 
-hack/tools/tool/logcheck/pkg/logcheck/testdata/src/github.com/go-logr/logr
-hack/tools/tool/logcheck/pkg/logcheck/testdata/src/sigs.k8s.io/controller-runtime/pkg/log
+hack/tools/logcheck/pkg/logcheck/testdata/src/github.com/go-logr/logr
+hack/tools/logcheck/pkg/logcheck/testdata/src/sigs.k8s.io/controller-runtime/pkg/log
 
 .golangci.yaml
 # Virtual go & fuse
