+++ .golangci.yaml.in
@@ -270,7 +270,7 @@ linters:
       logcheck:
         path: <<LOGCHECK_PLUGIN_PATH>>/logcheck.so
         description: Check structured logging calls to logr.Logger instances
-        original-url: github.com/gardener/gardener/hack/tools/tool/logcheck
+        original-url: github.com/gardener/gardener/hack/tools/logcheck
   exclusions:
     generated: lax
     rules:
