+++ Makefile
@@ -107,14 +107,14 @@ docker-push:
 #####################################################################
 
 PKG_APIS_DIR  := $(REPO_ROOT)/pkg/apis
-TOOLS_MOD_DIR := $(REPO_ROOT)/hack/tools/tool
+TOOLS_MOD_DIR := $(REPO_ROOT)/hack/tools
 LOGCHECK_DIR  := $(TOOLS_MOD_DIR)/logcheck
 
 .PHONY: tidy
 tidy:
 	@go work use
 	@go work sync
-	@GO111MODULE=on go mod tidy
+	@go mod tidy
 	@cd $(PKG_APIS_DIR); go mod tidy
 	@cd $(TOOLS_MOD_DIR); go mod tidy
