+++ cmd/gardenadm/main.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-//go:generate go run ../../hack/tools/cli-reference-generator -O ../../docs/cli-reference/gardenadm gardenadm
+//go:generate go run ../../hack/generators/cli-reference-generator -O ../../docs/cli-reference/gardenadm gardenadm
 
 package main
