+++ go.work
@@ -2,6 +2,6 @@ go 1.26.0
 
 use (
 	./
-	./hack/tools/tool
+	./hack/tools
 	./pkg/apis
 )
