+++ hack/generators/cli-reference-generator/main.go
