+++ hack/generators/oidcmeta/main.go
