+++ hack/tools/logcheck/pkg/logcheck/logcheck.go
