+++ hack/tools/logcheck/pkg/logcheck/testdata/src/no-logr/no_logr.go
