+++ hack/tools/logcheck/pkg/logcheck/testdata/src/use-logr/helper/funcs.go
