+++ hack/tools/logcheck/pkg/logcheck/testdata/src/use-logr/use_logr.go
