--- hack/tools/tool/doc.go
@@ -1,8 +1,0 @@
-// SPDX-FileCopyrightText: SAP SE or an SAP affiliate company and Gardener contributors
-//
-// SPDX-License-Identifier: Apache-2.0
-
-package tool
-
-// Import controller-runtime to get version from this go.mod (synced via go.work).
-import _ "sigs.k8s.io/controller-runtime"
